refactored browse page
