+++ client/src/components/index.js
@@ -2,7 +2,7 @@ import Header from './Header'
 import Footer from './Footer'
 import SearchBar from './SearchBar/Search'
 import Login from './Login/index'
-import Category from './Category/index'
+import Category from './TutorialCard/index'
 import CommentInput from './CommentInput'
 import TutorialInput from './TutorialInput'
 import StepDisplay from './StepDisplay'
